Skip some Lean methods overloads (#55)
* Skip some Lean methods overloads

Avoid generating stubs for some methods overloads so that users can get autocompletion and don't have to implement fully typed methods when there are multiple overloads.

This is because when there are @overloads in the stubs, the actual method implementation must be typed including all of the parameters and return types that are handled by every single overload.

* Ignore members with StubsIgnore attribute

* Trigger CI

* Revert some changes